It seems that in ExnerFoamCP, theta and thetaf can evolve independently. Nowhere in ExnerFoamCP is theta reset from thetaf or from b or bf. This needs to be done. theta should not be remembered from one time-step to the next.
In exnerFoamCP, b is now contructed with a name

I have put an option on setTheta:
setTheta -CP
for Charney-Phillips staggering so that setTheta recontructs theta and thetaf on vertical faces from thetaf on horizontal faces
It seems that in ExnerFoamCP, theta and thetaf can evolve independently. Nowhere in ExnerFoamCP is theta reset from thetaf or from b or bf. This needs to be done. theta should not be remembered from one time-step to the next.

In exnerFoamCP, b is now contructed with a name

I have put an option on setTheta:
setTheta -CP
for Charney-Phillips staggering so that setTheta recontructs theta and thetaf on vertical faces from thetaf on horizontal faces