cylinderMesh
Wraps up a 2d mesh in the x-z plane into a cylinder with axis in the z direction.
stitchMesh is then needed to stitch together the left and right boundaries

scalarDeformationFoam
A version of advectionFoam which updates the velocity field at every time-step for Yumeng's planar version of Lauritzen's deformational advection test case

gmtFoam
Added functionality for plotting data on cylidrical meshes. projection2D can be set to cylinder

FitData
Modified the polynomial fit schemes so that it uses a 2d fit on a curved 2d geometry

setGaussians
Initial conditions of a scalar consisting of the sum of a number of Gaussians. This initial condition application has some good aspects to its design. It would be worth generalising this code to use other functions than Gaussians using auto pointers and run-time selection tables

run/deformingAdvection
An example test case of scalar deformation in a cylindrical geometry