Use stencil geometry instead of proc-local fvMesh
The stencil geometry uses the global mesh, not the processor-local mesh,
so we should use it in findFaceDirs to establish the stencil-local
coordinates.

This changed has stopped a divide-by-zero SIGFPE when normalising a
zero-vector jdir, but results are still different between the serial and
parallel runs, so there are still more bugs lurking.