I am looking at a case of James's which uses the compressible Euler solve exnerFoamH. (The case is not currently in a repo). The case crashes after ~10000 seconds due to an instability at the outlet boundary (top corner). I thought that I would put my debugging notes here so that others can read/comment.
James said that this case ran fine a year ago but no longer runs stably.

The dependent variables are Exner, theta and face normal winds. The boundaries consist of an inlet, an outlet, the ground and a rigid lid (2D). The boundary conditions are:

inlet:
velocity is (10,0,0)
theta is fixed, stratified value
Exner is fixedFluxBuoyantExner (grad(Exner) = g/(cp theta))

outlet
velocity is not specified
theta is zeroGradient
Exner is fixedFluxBuoyantExner

top
velocity should be no normal flow - check this!
theta is fixedValue (391.024758006K)
Exner is fixedFluxBuoyantExner

ground
velocity should be no normal flow - check this!
theta is fixedGradient
Exner is fixedFluxBuoyantExner

Observation - this is an open domain and the pressure is not fixed anywhere. This could be a problem. One solution might be to fix Exner at the outlet boundary

Observation - the convergence is not always good. There are 2 pressure solves per time-step and often the initial residual only goes from ~9e-5 to 2e-6. This convergence is not really enough. We really need more convergence per iteration.

Observation - the consistency between the density from the continuity equation and the density calculated from Exner from the equation of state is not very good. The maximum difference fraction is around 0.000591225497679 (but grows when the model begins to crash)

Easy thing to check - 3 iterations per time-step
Change nOuterCorrectors to 3 in system/fvSolution
This doesn't help Maximum rho difference fraction.
And it still crashes at the same time (tiny bit before actually) with Exner going negative in the TR corner and the flow acceleration towards this outlet corner.
It also gets cold in this corner

The consistency between rho from continuity and rho from Exner really needs to be better but before examining the code, I will just try fixing Exner at the outlet by setting the writeFormat to ascii in system/controlDict and manually editing 0/Exner after running setExnerBalancedH.
That doesn't help "Maximum rho difference fraction"
but does improve convergence each time-step. The initial residual now goes down more than a factor of 10 between the first and second iteration - good.
There is now no hint of instability at outlet boundary after 10000 seconds and everything still looks fine at the end of the run (18000 seconds).

This presumably is not the same as the code that worked a few years ago. I don't know why that one ran and this one doesn't.

Next need to look at the rho inconsistency problem.

Compare ExnerFoamH with partitionedExnerFoam/extensiveForm as I have done lots of work on the consistency of partitionedExnerFoam.
On the face of it the two implementations look very similar. Check for bugs in the time-stepping by setting
offCentre 1;
SIgravityWaves false;
in system/fvSchemes
That doesn't make any difference. Revert

Not sure what to do. In exnerEqn.H try replacing
U += ExnerEqn.flux()/offCentre;
with
U -= gradPcoeff*fvc::snGrad(Exner)*mesh.magSf();
no difference. Good

Have a look at where the differneces are. Eg, are they just at boundaries or are they where the mesh is non-orthogonal?
Write out rhoDiff every time-step and plot
The internal differnces are max ~5e-9 so this is a boundary problem.
Differences at the inlet are ~1e-6
Differences at the outlet are ~1e-9
Differences on the gound are up to 1e-5
Differences at the top BC are ~1e-6
But these are absolute differences. Look at normalised differences.
Internal normalised differences are again tiny. Look at boundaries.
Differences at the inlet are up to 1e-4
Differences at the outlet are up to 1e-7
Differences on the gound are up to 1e-5
Differences at the top BC are ~4e-4
Things to try:
Make sure that Psi and divU are set on boundaries
That didn't help :-(f
Only print out the rho inconsistency for internal values. This will then only pick up problems rather than picking up boundary problems that I don't know how to fix.
Yup, that made the problem go away - bit of a cheat