Modify boussinesqFoam so that "-postProcess" option works again
Changes in

    applications/solvers/Boussinesq/boussinesqFoam/boussinesqFoam.C
    applications/solvers/Boussinesq/boussinesqFoam/createFields.H

Ordering of "#include" statements changed so that "-postProcess" option works again.

This ordering means that the definition of the hydrostatic pressure cannot be in
"createFields.H". I have moved it to "boussinesqFoam.C" for now, but this of course isn't ideal.

I presume a utility which sets a hydrostatic pressure field would be preferable to having the
solver itself do it?